PLANET-7898: Hide the Remember checkbox on login screen (#2742)
- Hide the checkbox
- Add a new class
- Move related functions
